tests/wpt/mozilla/tests/mozilla/css-transition-cancel-event.html

<!doctype html>
<html>
<head>
<meta charset=utf-8>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

</head>
<body>

<script>
const loaded = new Promise(resolve => {
  addEventListener('load', () => {
    setTimeout(resolve, 100);
  });
});

promise_test(async t => {
  await loaded;

  const div1 = document.createElement('div');
  const div2 = document.createElement('div');
  
  div1.setAttribute('style', 'transition: all .01s linear; ' +
                             'padding-left: 1px');
  div2.setAttribute('style', 'transition: all 50s linear; ' +
                             'padding-left: 1px');
  document.body.appendChild(div1);
  document.body.appendChild(div2);

  getComputedStyle(div1).paddingLeft;
  getComputedStyle(div2).paddingLeft;
  div1.style.paddingLeft = '10px';
  div2.style.paddingLeft = '10px';

  div1.ontransitionend = () => {
    // Cancel `div2`'s transition
    div2.remove();
  };

  const watcher = new EventWatcher(t, div2, [ 'transitioncancel' ]);

  // Should receive `transitioncancel` soon
  // Don't rely on timeout because it would unblock the event loop and
  // trigger `transitioncancel`. We want the test to pass only if
  // `transitioncancel` is delivered soon enough.
  const start = Date.now();
  await watcher.wait_for('transitioncancel');
  
  assert_less_than(Date.now() - start, 1000);
}, 'transitioncancel is delivered promptly');

</script>
</body>
</html>
